refactor: removed magic numbers, fixed API requests imports
