+++ frontend/src/api/courses/getPaginatedCourses.ts
@@ -7,6 +7,9 @@ import { API } from 'constants/routes';
 import { errorSnackbar } from 'constants/snackbarVariant';
 import { Course } from 'types/course';
 
+const PAGE_CHANGE = 1;
+const FIRST_PAGE = 1;
+const EMPTY_LENGTH = 0;
 interface HookResult {
   fetchNextPage: () => void;
   isLoading: boolean;
@@ -32,8 +35,10 @@ const useGetPaginatedCourses = (): HookResult => {
     'paginatedCoursesList',
     getCourses,
     {
-      getPreviousPageParam: (firstPage) => (firstPage.page === 1 ? false : firstPage.page - 1),
-      getNextPageParam: (lastPage) => (lastPage.courses.length > 0 ? lastPage.page + 1 : false),
+      getPreviousPageParam: (firstPage) =>
+        firstPage.page === FIRST_PAGE ? false : firstPage.page - PAGE_CHANGE,
+      getNextPageParam: (lastPage) =>
+        lastPage.courses.length > EMPTY_LENGTH ? lastPage.page + PAGE_CHANGE : false,
       refetchOnWindowFocus: false,
       onError: handleSubmitError,
     },
