+++ frontend/src/api/myCourses/index.ts
@@ -3,3 +3,4 @@ export { default as useFinishClientCourse } from './finishClientCourse';
 export { default as useGetClientCourseInfo } from './getClientCourseInfo';
 export { default as usePassClientCourse } from './passClientCourse';
 export { default as searchClientCourses } from './searchClientCourses';
+export { default as useGetClientPaginatedCourses } from './getClientPaginatedCourses';
